Removed a few PHP warnings
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/formhandler/trunk@49894 735d13b6-9817-0410-8766-e36946ffe9aa

+++ Resources/PHP/Hooks/class.tx_formhandler_clearCache.php
@@ -3,7 +3,9 @@
 class tx_formhandler_clearCache {
 
 	public function clearCache($params) {
-		unlink(PATH_site . 'typo3temp/' . 'formhandlerClassesCache.txt');
+		if(file_exists(PATH_site . 'typo3temp/' . 'formhandlerClassesCache.txt')) {
+			unlink(PATH_site . 'typo3temp/' . 'formhandlerClassesCache.txt');
+		}
 	}
 
 }
